+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/TransferDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class TransferDto {
 
   private Long id;
-  private String transferDate;
+  private String transferTime;
   private Long senderLabId;
   private String senderLabLabel;
   private Long senderGroupId;
@@ -23,12 +23,12 @@ public class TransferDto {
     this.id = id;
   }
 
-  public String getTransferDate() {
-    return transferDate;
+  public String getTransferTime() {
+    return transferTime;
   }
 
-  public void setTransferDate(String transferDate) {
-    this.transferDate = transferDate;
+  public void setTransferTime(String transferTime) {
+    this.transferTime = transferTime;
   }
 
   public Long getSenderLabId() {