+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/page/AbstractListPage.java
@@ -119,7 +119,7 @@ public interface AbstractListPage {
     public static final String TARGETED_SEQUENCING_REQD = "Targeted Sequencing Required";
     public static final String TISSUE_ORIGIN = "Tissue Origin";
     public static final String TISSUE_TYPE = "Tissue Type";
-    public static final String TRANSFER_DATE = "Transfer Date";
+    public static final String TRANSFER_TIME = "Transfer Time";
     public static final String TYPE = "Type";
     public static final String UNIQUE_DUAL_INDICES = "Unique Dual Indices";
     public static final String UNKNOWN = "Unknown";