+++ miso-web/src/main/webapp/scripts/list_itemtransfer.js
@@ -46,8 +46,8 @@ ListTarget.itemtransfer = (function() {
         include: true,
         bSortable: false
       }, {
-        sTitle: 'Transfer Date',
-        mData: 'transferDate',
+        sTitle: 'Transfer Time',
+        mData: 'transferTime',
         include: true,
         iSortPriority: 1
       }, {