+++ sqlstore/src/main/resources/db/migration/V8505__transfer_time.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Transfer CHANGE COLUMN transferDate transferTime DATETIME NOT NULL;